requirements.txt
python-telegram-bot>=20.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
aiohttp>=3.8.0